[Silver V] Title: 최소공배수, Time: 0 ms, Memory: 1112 KB -BaekjoonHub

+++ 백준/Silver/13241. 최소공배수/최소공배수.c
@@ -4,9 +4,9 @@
 int gcd(long int, long int);
 
 int main() {
-    long long int a, b;
-    if (scanf("%lld %lld", &a, &b) != 2) exit(0);
-    printf("%lld\n", (a*b)/gcd(a,b));
+    long int a, b;
+    if (scanf("%ld %ld", &a, &b) != 2) exit(0);
+    printf("%ld\n", (a*b)/gcd(a,b));
     return 0;
 }
 int gcd(long int a, long int b) {
